+++ packages/gcf-utils/CHANGELOG.md
@@ -87,6 +87,31 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
+## [12.0.0](https://www.github.com/googleapis/repo-automation-bots/compare/gcf-utils-v11.4.3...gcf-utils-v12.0.0) (2021-07-09)
+
+
+### ⚠ BREAKING CHANGES
+
+* **gcf-utils:** upgrade to probot 12 (#2253)
+
+### Bug Fixes
+
+* **gcf-utils:** upgrade to probot 12 ([#2253](https://www.github.com/googleapis/repo-automation-bots/issues/2253)) ([57bc7cf](https://www.github.com/googleapis/repo-automation-bots/commit/57bc7cfde2f3d23cb1bd263310b5c131e6c04269))
+
+### [11.4.3](https://www.github.com/googleapis/repo-automation-bots/compare/gcf-utils-v11.4.2...gcf-utils-v11.4.3) (2021-07-07)
+
+
+### Bug Fixes
+
+* return a 400 on requests with missing or bad signatures ([#2240](https://www.github.com/googleapis/repo-automation-bots/issues/2240)) ([6cf6964](https://www.github.com/googleapis/repo-automation-bots/commit/6cf696442cd69030d5b5bda7ddce3dee846e5fef))
+
+### [11.4.2](https://www.github.com/googleapis/repo-automation-bots/compare/gcf-utils-v11.4.1...gcf-utils-v11.4.2) (2021-07-02)
+
+
+### Bug Fixes
+
+* detect default task target runtime from environment ([#2233](https://www.github.com/googleapis/repo-automation-bots/issues/2233)) ([7f262e8](https://www.github.com/googleapis/repo-automation-bots/commit/7f262e8f3208a7f5247284fb2880da0e8ebeb359)), closes [#2194](https://www.github.com/googleapis/repo-automation-bots/issues/2194)
+
 ### [11.4.1](https://www.github.com/googleapis/repo-automation-bots/compare/gcf-utils-v11.4.0...gcf-utils-v11.4.1) (2021-06-30)